refactor: do not use cli text args directly in lib
